# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OCTAGON - Local-first operator assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List workflows
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, active, completed, archived]
          required: false
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get a workflow with steps
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Workflow with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDetail"
        "404":
          description: Not found
    patch:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update a workflow
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "204":
          description: Deleted
  /workflows/{id}/steps:
    post:
      operationId: addWorkflowStep
      tags: [workflows]
      summary: Add a step to a workflow
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowStepInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStep"
  /workflows/{id}/steps/{stepId}:
    patch:
      operationId: updateWorkflowStep
      tags: [workflows]
      summary: Update a workflow step status
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: path
          name: stepId
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowStepUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStep"
...
components:
  schemas:
...
    Workflow:
      type: object
      required: [id, name, description, status, stepCount, completedSteps, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, active, completed, archived]
        stepCount:
          type: integer
        completedSteps:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowDetail:
      type: object
      required: [id, name, description, status, steps, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, active, completed, archived]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    WorkflowUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, active, completed, archived]
        tags:
          type: array
          items:
            type: string
    WorkflowStep:
      type: object
      required: [id, workflowId, order, action, status]
      properties:
        id:
          type: integer
        workflowId:
          type: integer
        order:
          type: integer
        action:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        result:
          type: ["string", "null"]
        executionMode:
          type: string
          enum: [local, cloud, hybrid]
    WorkflowStepInput:
      type: object
      required: [action, order]
      properties:
        action:
          type: string
          minLength: 1
        description:
          type: string
        order:
          type: integer
        executionMode:
          type: string
          enum: [local, cloud, hybrid]
    WorkflowStepUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        result:
          type: string
        executionMode:
          type: string
          enum: [local, cloud, hybrid]
...