# lib/api-zod/src/generated/types/workflow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OCTAGON - Local-first operator assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowStatus } from './workflowStatus';

export interface Workflow {
  id: number;
  name: string;
  description: string;
  status: WorkflowStatus;
  stepCount: number;
  completedSteps: number;
  tags?: string[];
  createdAt: Date;
  updatedAt: Date;
}