# lib/api-zod/src/generated/types/workflowDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OCTAGON - Local-first operator assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowDetailStatus } from './workflowDetailStatus';
import type { WorkflowStep } from './workflowStep';

export interface WorkflowDetail {
  id: number;
  name: string;
  description: string;
  status: WorkflowDetailStatus;
  steps: WorkflowStep[];
  tags?: string[];
  createdAt: Date;
  updatedAt: Date;
}