# lib/api-zod/src/generated/types/workflowStepInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OCTAGON - Local-first operator assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowStepInputExecutionMode } from './workflowStepInputExecutionMode';

export interface WorkflowStepInput {
  /** @minLength 1 */
  action: string;
  description?: string;
  order: number;
  executionMode?: WorkflowStepInputExecutionMode;
}